demo/requirements.txt
Django==2.0.2
django-ajax-selects==1.7.0
pytz==2018.3
